+++ Assets/Scripts/FrameCounter.cs
@@ -7,7 +7,7 @@ public class FrameCounter : MonoSingleton<FrameCounter>
 	// The basic tick that doesn't try to be before/after anything else.
 	public EventHandler OnAwake;
 	public EventHandler OnStart;
-    public EventHandler<int, float> OnUpdate;
+    public EventHandler<long, float> OnUpdate;
 
 	public int count { get; private set; }
     public float deltaTime {
