+++ Assets/Scripts/PlayerInputControllerGamepad.cs
@@ -2,7 +2,7 @@
 
 public class PlayerInputControllerGamepad : PlayerInputController
 {
-	public override void HandleUpdate(int currentFrame, float deltaTime)
+	public override void HandleUpdate(long currentFrame, float deltaTime)
     {
 		base.HandleUpdate(currentFrame, deltaTime);
